Added Random Champion Command
Currently have a working framework for this command. Will need to update
it with something more elegant when I get time. Also started the work
for summoner information. This version is more of a backup of what I've
already done. (It runs fine, its just not as dynamic or elegant as I'd
personally like).